% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firthDat.R
\docType{data}
\name{firthDat}
\alias{firthDat}
\title{firthDat}
\format{An object of class \code{"data.frame"}}
\usage{
data(firthDat)
}
\description{
a generated data
}
\examples{

 data(firthDat)

 names(firthDat)


}
\keyword{datasets}
