% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefirth.R
\name{geefirth}
\alias{geefirth}
\title{Fitting Firth-type GEE for  correlated binary data with separation or near-to-separation}
\usage{
geefirth(formula, id = id, data = parent.env(),
  corstr = "independence")
}
\arguments{
\item{formula}{Similar as geeglm}

\item{id}{cluster variable and requirs to be sorted}

\item{data}{data frame that contains the variables mentioned in id and formula}

\item{corstr}{working correlation structure, "independence" (default), "ar1", "unstr", "exchangeable".}
}
\description{
geefirth fits GEE with Firth-type penalization to provide
bias-corrected finite  estimate of the regression coefficient  in case of separation or near-to-separation. In addition,
it provides bias-correctedsandwich estimate of the standard error of the penalized GEE estimate.
}
\examples{

data(quasiDat)

# Fitting GEE for quasi-separated data

 geefirth(y ~ x + obstime, id=id,
       data=quasiDat, corstr = "exchangeable");

data(nearDat)

# Fitting GEE for near to quasi-separated data

 geefirth(y ~ x + obstime, id=id,
          data=nearDat, corstr = "exchangeable");

}
\author{
Momenul Haque Mondol \email{mmondol@isrt.ac.bd}, M. Shafiqur Rahman \email{
shafiq@isrt.ac.bd}
}
