% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefirth_data.R
\docType{data}
\name{geefirth_data}
\alias{geefirth_data}
\title{Random longitudinal binary data}
\format{data frame}
\usage{
geefirth_data
}
\description{
. It has two data sets. One is "quasi_sep" (having quasi-complete separation problem)
and another "near_to_sep" (having near to separation problem). Each data contains one binary and one
 time (continuous) covariate.It is generated from true exchangeable correlation with icc 0.5, prevelence
of event in binary covariate 0.3 and  log odds ratio -2.5, 1.6 and 0.2
for intercept, x and obstime respectively.
}
\keyword{datasets}
