% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefirthm.R
\docType{package}
\name{geefirthm}
\alias{geefirthm}
\alias{geefirthm-package}
\title{Fitting Firth-type GEE for  correlated binary data with separation or near-to-separation}
\description{
geefirth fits GEE with Firth-type penalization to provide
bias-corrected finite  estimate of the regression coefficient  in case of separation or near-to-separation. In addition,
it provides bias-correctedsandwich estimate of the standard error of the penalized GEE estimate.
}
\author{
Momenul Haque Mondol \email{mmondol@isrt.ac.bd}
}
