% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearDat.R
\docType{data}
\name{nearDat}
\alias{nearDat}
\title{nearDat}
\format{An object of class \code{"data.frame"}}
\usage{
data(nearDat)
}
\description{
a generated data
}
\examples{

 data(nearDat)

 names(nearDat)


}
\keyword{datasets}
