% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiDat.R
\docType{data}
\name{quasiDat}
\alias{quasiDat}
\title{a generated data}
\format{An object of class \code{"data.frame"}}
\usage{
data(quasiDat)
}
\description{
a generated data
}
\examples{

 data(quasiDat)

 names(quasiDat)


}
\keyword{datasets}
