% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geefirth.R
\name{geefirth}
\alias{geefirth}
\title{Fitting Firth-type GEE for  correlated binary data with separation or near-to-separation}
\usage{
geefirth(formula = formula(data), id = id, data = parent.frame(),
  corstr = "independence")
}
\arguments{
\item{formula}{Similar as geeglm}

\item{id}{cluster variable and requirs to be sorted}

\item{corstr}{working correlation structure, "independence" (default), "ar1", "unstr", "exchangeable".}
}
\description{
geefirth fits GEE with Firth-type penalization to provide
bias-corrected finite  estimate of the regression coefficient  in case of separation or near-to-separation. In addition,
it provides bias-correctedsandwich estimate of the standard error of the penalized GEE estimate.
}
\examples{

# loading data
data(geefirth_data)

# Fitting GEE for quasi-separated data

geefirth(y ~ x + obstime, id=quasi_sep$id, data=quasi_sep, corstr = "exchangeable");

# Fitting GEE for near to quasi-separated data

geefirth(y ~ x + obstime, id=near_to_sep$id, data=near_to_sep, corstr = "exchangeable");

}
\author{
Momenul Haque Mondol \email{mmondol@isrt.ac.bd}, M. Shafiqur Rahman \email{
shafiq@isrt.ac.bd}
}
